+++ src/main/java/nextstep/security/util/Base64Convertor.java
@@ -1,8 +1,13 @@
-package nextstep.security;
+package nextstep.security.util;
 
 import java.util.Base64;
 
-public class Base64Convertor {
+public final class Base64Convertor {
+
+    private Base64Convertor() {
+        throw new AssertionError();
+    }
+
     public static String encode(String value) {
         return Base64.getEncoder().encodeToString(value.getBytes());
     }
@@ -10,4 +15,5 @@ public class Base64Convertor {
     public static String decode(String value) {
         return new String(Base64.getDecoder().decode(value));
     }
+
 }
